IBX-7935: Refactored code
