+++ src/lib/Event/UserUpdateFieldOptionsEvent.php
@@ -12,24 +12,14 @@ use Ibexa\Contracts\ContentForms\Data\Content\FieldData;
 use Ibexa\Contracts\Core\Repository\Values\Content\Content;
 use Ibexa\Contracts\Core\Repository\Values\User\UserUpdateStruct;
 use Symfony\Component\Form\FormInterface;
-use Symfony\Contracts\EventDispatcher\Event;
 
-final class UserUpdateFieldOptionsEvent extends Event
+final class UserUpdateFieldOptionsEvent extends UserStructFieldOptionsEvent
 {
     /** @var \Ibexa\Contracts\Core\Repository\Values\Content\Content */
     private $content;
 
     /** @var \Ibexa\Contracts\Core\Repository\Values\User\UserUpdateStruct */
     private $userUpdateStruct;
-
-    /** @var \Symfony\Component\Form\FormInterface */
-    private $parentForm;
-
-    /** @var \Ibexa\Contracts\ContentForms\Data\Content\FieldData */
-    private $fieldData;
-
-    /** @var array<string, mixed> */
-    private $options;
 
     public function __construct(
         Content $content,
@@ -40,9 +30,8 @@ final class UserUpdateFieldOptionsEvent extends Event
     ) {
         $this->content = $content;
         $this->userUpdateStruct = $userUpdateStruct;
-        $this->parentForm = $parentForm;
-        $this->fieldData = $fieldData;
-        $this->options = $options;
+
+        parent::__construct($parentForm, $fieldData, $options);
     }
 
     public function getContent(): Content
@@ -50,49 +39,8 @@ final class UserUpdateFieldOptionsEvent extends Event
         return $this->content;
     }
 
-    public function setContent(Content $content): void
-    {
-        $this->content = $content;
-    }
-
     public function getUserUpdateStruct(): UserUpdateStruct
     {
         return $this->userUpdateStruct;
-    }
-
-    public function getParentForm(): FormInterface
-    {
-        return $this->parentForm;
-    }
-
-    public function getFieldData(): FieldData
-    {
-        return $this->fieldData;
-    }
-
-    /**
-     * @return array<string, mixed>
-     */
-    public function getOptions(): array
-    {
-        return $this->options;
-    }
-
-    /**
-     * @param array<string, mixed> $options
-     */
-    public function setOptions(array $options): void
-    {
-        $this->options = $options;
-    }
-
-    public function setOption(string $option, $value): void
-    {
-        $this->options[$option] = $value;
-    }
-
-    public function getOption(string $option)
-    {
-        return $this->options[$option] ?? null;
     }
 }
